# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "CategoryOut": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "image_url"
        ],
        "title": "CategoryOut",
        "type": "object"
      },
      "CategoryRef": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug"
        ],
        "title": "CategoryRef",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page_ProductCardOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProductCardOut"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[ProductCardOut]",
        "type": "object"
      },
      "ProductCardOut": {
        "properties": {
          "discount_percent": {
            "title": "Discount Percent",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "is_available": {
            "title": "Is Available",
            "type": "boolean"
          },
          "max_per_order": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Per Order"
          },
          "mrp_paise": {
            "title": "Mrp Paise",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "price_paise": {
            "title": "Price Paise",
            "type": "integer"
          },
          "stock_hint": {
            "enum": [
              "IN_STOCK",
              "LOW",
              "OUT"
            ],
            "title": "Stock Hint",
            "type": "string"
          },
          "unit_label": {
            "title": "Unit Label",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "unit_label",
          "price_paise",
          "mrp_paise",
          "discount_percent",
          "image_url",
          "is_available",
          "stock_hint",
          "max_per_order"
        ],
        "title": "ProductCardOut",
        "type": "object"
      },
...
      "ProductDetailOut": {
        "properties": {
          "category": {
            "$ref": "#/components/schemas/CategoryRef"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "discount_percent": {
            "title": "Discount Percent",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "is_available": {
            "title": "Is Available",
            "type": "boolean"
          },
          "max_per_order": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Per Order"
          },
          "mrp_paise": {
            "title": "Mrp Paise",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "price_paise": {
            "title": "Price Paise",
            "type": "integer"
          },
          "related": {
            "items": {
              "$ref": "#/components/schemas/ProductCardOut"
            },
            "title": "Related",
            "type": "array"
          },
          "stock_hint": {
            "enum": [
              "IN_STOCK",
              "LOW",
              "OUT"
            ],
            "title": "Stock Hint",
            "type": "string"
          },
          "unit_label": {
            "title": "Unit Label",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "unit_label",
          "price_paise",
          "mrp_paise",
          "discount_percent",
          "image_url",
          "is_available",
          "stock_hint",
          "max_per_order",
          "description",
          "category",
          "related"
        ],
        "title": "ProductDetailOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Bada Bazar API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/categories/{slug}": {
      "get": {
        "operationId": "get_category_api_v1_categories__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Category",
        "tags": [
          "catalog"
        ]
      }
    },
...
    "/api/v1/products": {
      "get": {
        "operationId": "list_products_api_v1_products_get",
        "parameters": [
          {
            "in": "query",
            "name": "category_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category Id"
            }
          },
          {
            "in": "query",
            "name": "in_stock_only",
            "required": false,
            "schema": {
              "default": false,
              "title": "In Stock Only",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "default",
              "enum": [
                "default",
                "price_asc",
                "price_desc"
              ],
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ProductCardOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Products",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/products/{product_id}": {
      "get": {
        "operationId": "get_product_api_v1_products__product_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Product Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Product",
        "tags": [
          "catalog"
        ]
      }
    },
...
  }
}